sched/{core,fair}: fix build for !SMP systems
Utilization is tracked only for !SMP systems, thus all related functions
cannot be used on unicore builds.

Currently, boosted_cpu_util is defined for !SMP systems to return
cpu_util_freq, which however is undefined for those systems.

This patch fixes that by removing the wrong definition of boosted_cpu_util
while ensuring that that API is never used on !SMP builds.

The only clients of this API are:
- update_capacity_of
  which is already empty defined for !SMP systems
- cpufreq_schedutil.c::sugov_get_util
  which is not available for !SMP systems
- sched_freq_tick_pelt

Let's fix the last client by making it an empty definition on !SMP systems.

Change-Id: Ifc4a9a15042684ee8db9caaa046fe338161bb053
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>